+++ header.php
@@ -4,7 +4,7 @@
  *
  * @package    Auberge
  * @copyright  2014 WebMan - Oliver Juhas
- * @version    1.0
+ * @version    1.2
  */
 
 
@@ -17,8 +17,7 @@
 
 ?>
 
-<!--[if IE 9]><html class="ie ie9 lie9 no-js" <?php language_attributes(); ?>><![endif]-->
-<!--[if (gt IE 9)|!(IE)]><!--><html class="no-js" <?php language_attributes(); ?>><!--<![endif]-->
+<html class="no-js" <?php language_attributes(); ?>>
 
 <head>
 
@@ -29,14 +28,6 @@
 	 */
 
 	wmhook_head_top();
-
-	if ( ! function_exists( '_wp_render_title_tag' ) ) :
-	?>
-
-	<title><?php wp_title( '|', true, 'right' ); ?></title>
-
-	<?php
-	endif;
 
 	wmhook_head_bottom();