+++ js/customizer-preview.js
@@ -3,14 +3,20 @@
  *
  * @package    Auberge
  * @copyright  2014 WebMan - Oliver Juhas
- * @version    1.0
+ *
+ * @since    1.0
+ * @version  1.2
  */
 
 
 
 
 
-jQuery( function() {
+( function( $ ) {
+
+
+
+	"use strict";
 
 
 
@@ -21,7 +27,7 @@ jQuery( function() {
 		wp.customize( 'blogname', function( value ) {
 
 			value.bind( function( to ) {
-				jQuery( '.site-title a' ).text( to );
+				$( '.site-title a' ).text( to );
 			} );
 
 		} ); // /blogname
@@ -29,11 +35,11 @@ jQuery( function() {
 		wp.customize( 'blogdescription', function( value ) {
 
 			value.bind( function( to ) {
-				jQuery( '.site-description, .site-banner-header .highlight' ).text( to );
+				$( '.site-description, .site-banner-header .highlight' ).text( to );
 			} );
 
 		} ); // /blogdescription
 
 
 
-} );
+} )( jQuery );