+++ changelog.md
@@ -1,5 +1,51 @@
 # Auberge Changelog
 
+## 1.2
+
+* Added: Support for Beaver Builder page builder plugin
+* Added: Full compatibility with WordPress 4.1
+* Added: Demo content file download via online theme user manual
+* Added: Development versions of jQuery plugins
+* Added: Jetpack responsive videos support
+* Updated: Code improvements and optimization
+* Updated: Underscores updates implemented
+* Updated: Reorganized Customizer sections into panels
+* Updated: Updated hook names that weren't following naming convention
+* Updated: Removed obsolete functions and hooks
+* Updated: Removed demo content XML file from the theme folder
+* Updated: User manual updated
+* Updated: Improved styles
+* Updated: Updated localization
+* Updated: Visual editor addons improved
+* Updated: Updated JavaScript files
+* Updated: Added proper escaping for Customizer output CSS
+* Fixed: Hook names
+* Fixed: Submenu plus icon position
+* Fixed: Postlist article hover custom border color not applied
+* Fixed: Added link to access posts with no title
+
+#### Files changed:
+
+	archive.php
+	functions.php
+	header.php
+	sidebar-footer.php
+	css/_custom.css
+	inc/setup-theme-options.php
+	inc/setup.php
+	inc/beaver-builder/beaver-builder.php
+	inc/customizer/customizer.php
+	inc/jetpack/jetpack.php
+	inc/lib/admin.php
+	inc/lib/core.php
+	inc/lib/visual-editor.php
+	js/customizer-preview.js
+	js/scripts.js
+	js/skip-link-focus-fix.js
+	js/dev/imagesloaded.pkgd.js
+	js/dev/slick.js
+
+
 ## 1.1.5
 
 * Updated customizer to support WP4.1